Upgraded Python2 scripts
Upgraded Python2 scripts to Python 3 using 2to3 tool.
This is to mitigrate security risks of Python2 deprecation.

Tracked-On: OAM-90216

+++ src/gallium/auxiliary/util/u_format_pack.py
@@ -36,7 +36,7 @@
 '''
 
 
-from __future__ import division, print_function
+
 
 import sys
 
@@ -44,7 +44,7 @@ from u_format_parse import *
 
 
 if sys.version_info < (3, 0):
-    integer_types = (int, long)
+    integer_types = (int, int)
 
 else:
     integer_types = (int, )
